fix app 在 iPad 上 present UIActivityViewController 出错的问题。
see https://bugly.qq.com/v2/crash-reporting/crashes/10e24876f2/10042?pid=2
Your application has presented a UIActivityViewController (<UIActivityViewController: 0x102997200>). In its current trait environment, the modalPresentationStyle of a UIActivityViewController with this style is UIModalPresentationPopover. You must provide location information for this popover through the view controller's popoverPresentationController. You must provide either a sourceView and sourceRect or a barButtonItem. If this information is not known when you present the view controller, you may provide it in the UIPopoverPresentationControllerDelegate method -prepareForPopoverPresentation.